Refactor: Removing dart analysis warnings
Remoção de warnings do dart analysis
