+++ lib/features/data/datasources/farm_datasource/farm_datasource_impl.dart
@@ -41,6 +41,7 @@ class FarmDatasourceImpl with UriBuilder implements FarmDatasource {
     }
   }
 
+  @override
   Future<List<FarmModel>> getCostCenterFarmsById(String costCenterId) async {
     final String url = mountUrl(
       AppEndpoints.baseUrlProtocolWithSecurity,
