+++ lib/features/presenter/modules/service_order/create_service_order_page/createServiceOrderTabs/select_plots_page.dart
@@ -15,6 +15,7 @@ class SelectPlotsPage extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     final controller = Modular.get<CreateServiceOrderController>();
+    controller.getPlotsOptions();
     return Observer(
         builder: (context) =>
             Column(mainAxisAlignment: MainAxisAlignment.center, children: [
@@ -55,7 +56,7 @@ class SelectPlotsPage extends StatelessWidget {
                                     builder: (context) => SizedBox(
                                           height: controller.itemHeight,
                                           child: CustomCheckableListItemWidget(
-                                              indexIsChecked: controller
+                                              isChecked: controller
                                                   .selectedPlots
                                                   .contains(controller
                                                       .plotsOptions[index].id),
